e-sport-analytical-app/backend/requirements.txt
# E-Sport Analytics Platform - Backend Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
alembic>=1.12.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# External API Integration
aiohttp>=3.9.0
asyncio
requests>=2.31.0

# Azure Integration (Optional)
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# Caching & Performance
redis>=5.0.0  # Optional for production caching

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Optional: Production server
gunicorn>=21.0.0
